minor checks on the data.
1. *attendance* == *spettatori*;
2. We can cast *is_win* to a boolean

I gave a better look to the **Note.md** file and I think we can try to get out the json files.
All we have to do is to define a new library that can execute methods to generate the cleaned up json file for every table of the servers.